refactor, fuzz: Move SimualtionTest to fuzz/util
